domain map fixes
